feat: ajout de MyProfile.php et update du compte utilisateur
- Création de la page MyProfile et ajout de la fonctionnalité de modification
- Intégration du formulaire pour modifier le pseudo et le mot de passe
- Lien avec l'API (requête PUT) pour mettre à jour la base de données

+++ DWL_PP/siteweb/DWL.php
@@ -1,7 +1,12 @@
 <?php 
 require_once "config/configuration.php";
 require_once "service/MovieService.php";
-
+session_start();
+$username = $_SESSION['username'] ?? 'Unknown';
+if ($username == "Unknown") {
+    header("Location: Login.php");
+    exit;
+}
 if (isset($_POST["delete_id"])) {
     $id_a_supprimer = $_POST["delete_id"];
     MovieService::deleteMovieWL($id_a_supprimer);
